requirements.txt
pyperclip~=1.8.2
pystray~=0.19.3
Pillow~=9.1.0
py2exe~=0.11.1.0